refactor: host builder pipeline logging
This change moves implementations to multiple, more sensible components. It also configures SeriLog correctly in the host builder pipeline.
