+++ src/DaemonRunner/DaemonRunner/Service/Extensions/ServiceCollectionExtensions.cs
@@ -1,0 +1,15 @@
+﻿using Microsoft.Extensions.DependencyInjection;
+
+namespace NetDaemon.Service.Extensions
+{
+    public static class ServiceCollectionExtensions
+    {
+        public static IServiceCollection AddNetDaemon(this IServiceCollection services)
+        {
+            services.AddHttpClient();
+            services.AddHostedService<RunnerService>();
+
+            return services;
+        }
+    }
+}
